Try to make flexible project structure. Now, http server is a separate object that need to be launched. And also context is used. Made based on the example available at the link.
